requirements.txt
flask-restful
flask-mysql
cryptography
